fix(frontend): resolve remaining react-doctor warnings and PR review feedback
- Extract FileTable from FolderDetail to reduce component size (304→170 lines)
- Move useSearchParams into dedicated WithParams wrapper files for
  proper Suspense boundary detection
- Extract provider fetch into useCatalogDefaults hook
- Split app/page.tsx into server page with metadata + client component
- Fix Conversation.id types from number to string (backend returns UUID)
- Change nested nav to div in AppSidebar to avoid duplicate landmarks
- Replace role="button" on UploadMenu backdrop with aria-hidden
- Wrap setMessages in useCallback for stable identity
